GERANDO PRIMEIRO PEDIDO
Corrigido alguns detalhes.

Após gerar os 3 primeiros tempos para a chegada dos pedidos e incluídos
na lista ordenada, verifiquei qual era o menor tempo e atualizei o tempo
atual para o menor de todos os tempos, então naquele instante foi gerado
um ROLAMENTO e incluído em uma máquina. Após, foi gerado um novo tempo
de pedido e inserido novamente na lista ordenada.

+++ fabrica/listaordenada.c
@@ -2,7 +2,7 @@
 #include "listaordenada.h"
 #include "nodo.h"
 
-lista* criar_lista_ordenada(double (*func)(void*,void*)){
+lista* criar_lista_ordenada(int (*func)(void*,void*)){
 	lista *lst = malloc(sizeof(lista));
 	lst->func = func;
 	lst->tam = 0;
@@ -94,6 +94,7 @@ void* obter_ordenada(lista *lst, int pos){
 		for (int i = 0; i < pos; ++i) nTemp = nTemp->prox;
 		elem = get_ElemNodo_ordenada(nTemp);
 	}
+
 	return elem;
 }
